CP - Functional powerplant scripts standalone (untested through Snakemake)
